requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
TTS==0.22.0
torch==2.1.0
torchaudio==2.1.0
ffmpeg-python
boto3
requests
python-dotenv
